+++ src/main/java/com/sarida/logtown/service/AdminService.java
@@ -1,7 +1,11 @@
 package com.sarida.logtown.service;
 
 import com.sarida.logtown.dto.ApiResponseDto;
+import com.sarida.logtown.dto.CommentListResponseDto;
+import com.sarida.logtown.dto.PostListResponseDto;
 import com.sarida.logtown.dto.UserInfoListResponseDto;
+
+import java.util.List;
 
 public interface AdminService {
     /**
@@ -34,4 +38,18 @@ public interface AdminService {
      * @return 사용자 dto list
      */
     UserInfoListResponseDto getAllUserInfos();
+
+    /**
+     * 게시글 전체 보기
+     *
+     * @return 전체 게시글 list
+     */
+    PostListResponseDto getAllPosts();
+
+    /**
+     * 댓글 전체 보기
+     *
+     * @return 전체 댓글 list
+     */
+    CommentListResponseDto getAllComments();
 }
