+++ src/main/java/com/sarida/logtown/service/CommentService.java
@@ -1,7 +1,6 @@
 package com.sarida.logtown.service;
 
 import com.sarida.logtown.dto.ApiResponseDto;
-import com.sarida.logtown.dto.CommentListResponseDto;
 import com.sarida.logtown.dto.CommentRequestDto;
 import com.sarida.logtown.entity.Comment;
 import com.sarida.logtown.security.UserDetailsImpl;
@@ -54,10 +53,4 @@ public interface CommentService {
      */
     Comment findComment(Long commentId);
 
-    /**
-     * 모든 댓글 조회(관리자)
-     *
-     * @return 모든 댓글
-     */
-    CommentListResponseDto getAllComments();
 }
